requirements.txt
fastapi
uvicorn
streamlit
qdrant-client
langchain
langchain-community
langchain-qdrant
langchain-ollama
langgraph
pydantic
python-dotenv
python-multipart
fastembed
pypdf
PyPDF2
typing-extensions
groq 
langchain_groq